requirements.txt
tree-sitter>=0.22,<1.0
tree-sitter-python>=0.23,<1.0
ollama>=0.4
mlx-lm>=0.0.12
click>=8.0
pydantic>=2.0
pyyaml>=6.0
watchdog>=4.0
fastapi>=0.115
uvicorn>=0.34
